Moved the OIDC express logic out of the index file, will refactor that further in another commit
